refactor(report): split browser_evidence_compare.rs god-file into modules
